# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestión de Inventario",
        "description": "Documentación de la API con Swagger",
        "contact": {
            "email": "soporte@tuempresa.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost"
        },
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/alertas-vencimiento": {
            "post": {
                "tags": [
                    "AlertaVencimiento"
                ],
                "summary": "Crear una nueva alerta",
                "description": "Crear nueva alerta",
                "operationId": "4dd63d2d6ba2615fbcbd041abe2be742",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lote_id",
                                    "fecha_vencimiento",
                                    "estado"
                                ],
                                "properties": {
                                    "lote_id": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "fecha_vencimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-12-31"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "atendida"
                                        ],
                                        "example": "pendiente"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Alerta creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertaVencimiento"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/alertas-vencimiento/{id}": {
            "get": {
                "tags": [
                    "AlertaVencimiento"
                ],
                "summary": "Obtener una alerta específica",
                "description": "Ver una alerta por ID",
                "operationId": "4f81aa313e3d324c27ed3b81664eea09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertaVencimiento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "AlertaVencimiento"
                ],
                "summary": "Actualizar alerta por ID",
                "description": "Actualizar una alerta",
                "operationId": "7702b88e8c786cb92264a530b61b6c63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha_vencimiento": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-01-15"
                                    },
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "pendiente",
                                            "atendida"
                                        ],
                                        "example": "atendida"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alerta actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AlertaVencimiento"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "AlertaVencimiento"
                ],
                "summary": "Eliminar una alerta",
                "description": "Eliminar una alerta",
                "operationId": "e0dfc0b740cc99203b4ec3bb789888a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Alerta eliminada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Alerta eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Alerta no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AlertaVencimiento": {
                "required": [
                    "lote_id",
                    "fecha_vencimiento",
                    "estado"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lote_id": {
                        "type": "integer",
                        "example": 10
                    },
                    "fecha_vencimiento": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-12-31"
                    },
                    "fecha_alerta_generada": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-03-25T12:00:00Z"
                    },
                    "estado": {
                        "type": "string",
                        "enum": [
                            "pendiente",
                            "atendida"
                        ],
                        "example": "pendiente"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}